feat: add `spacemacs/view-current-eww`
This commit introduces a new function `spacemacs/view-current-eww` that allows users to view the URL of the current EWW buffer.

The function is bound to `SPC vv` in EWW mode, providing a convenient way to access the URL.

Additionally, a helper function `spacemacs//view-url` is introduced to handle URL dispatching, improving code organization and reusability.